WIP. Changed handle library to generic tool getter. Added flush handles main under cmd/
